Initial docker provider implementation
