+++ service/interface/interface.go
@@ -13,8 +13,9 @@ import (
 	. "github.com/logrusorgru/aurora"
 	"golang.org/x/term"
 
+	"github.com/pygmystack/pygmy/internal/runtimes/docker/containers"
+	"github.com/pygmystack/pygmy/internal/runtimes/docker/images"
 	"github.com/pygmystack/pygmy/service/color"
-	"github.com/pygmystack/pygmy/service/interface/docker"
 )
 
 // Setup will detect if the Service's image reference exists and will
@@ -26,15 +27,15 @@ func (Service *Service) Setup() error {
 	}
 
 	found := false
-	images, _ := docker.DockerImageList()
-	for _, image := range images {
+	imgs, _ := images.List()
+	for _, image := range imgs {
 		if strings.Contains(fmt.Sprint(image.RepoTags), Service.Config.Image) {
 			found = true
 		}
 	}
 
 	if !found {
-		if msg, err := docker.DockerPull(Service.Config.Image); err != nil {
+		if msg, err := images.Pull(Service.Config.Image); err != nil {
 			return err
 		} else if strings.Contains(msg, "already up to date") {
 			return fmt.Errorf(msg)
@@ -77,10 +78,10 @@ func (Service *Service) Start() error {
 	}
 
 	if purpose == "addkeys" {
-		if e := docker.DockerKill(name); e != nil {
+		if e := containers.Kill(name); e != nil {
 			fmt.Sprintln(e)
 		}
-		if e := docker.DockerRemove(name); e != nil {
+		if e := containers.Remove(name); e != nil {
 			fmt.Sprintln(e)
 		}
 		if e := Service.Create(); e != nil {
@@ -154,7 +155,7 @@ func (Service *Service) Status() (bool, error) {
 	if Service.HostConfig.AutoRemove {
 		return true, nil
 	}
-	containers, _ := docker.DockerContainerList()
+	containers, _ := containers.List()
 	for _, container := range containers {
 		for _, n := range container.Names {
 			if strings.Contains(n, name) && strings.HasPrefix(container.Status, "Up") {
@@ -170,7 +171,7 @@ func (Service *Service) Status() (bool, error) {
 // GetRunning will get a types.Container variable for a given running container
 // and it will not retrieve any information on containers that are not running.
 func (Service *Service) GetRunning() (types.Container, error) {
-	containers, _ := docker.DockerContainerList()
+	containers, _ := containers.List()
 	for _, container := range containers {
 		if _, ok := container.Labels["pygmy.name"]; ok {
 			if strings.Contains(container.Labels["pygmy.name"], Service.Config.Labels["pygmy.name"]) {
@@ -190,22 +191,22 @@ func (Service *Service) Clean() error {
 		return nil
 	}
 
-	Containers, _ := docker.DockerContainerList()
+	Containers, _ := containers.List()
 	for _, container := range Containers {
 		if container.Names[0] == name {
 			if pygmy {
 				name := strings.TrimLeft(container.Names[0], "/")
-				if e := docker.DockerKill(container.ID); e == nil {
+				if e := containers.Kill(container.ID); e == nil {
 					if !Service.HostConfig.AutoRemove {
 						color.Print(Green(fmt.Sprintf("Successfully killed %s\n", name)))
 					}
 				}
-				if e := docker.DockerStop(container.ID); e == nil {
+				if e := containers.Stop(container.ID); e == nil {
 					if !Service.HostConfig.AutoRemove {
 						color.Print(Green(fmt.Sprintf("Successfully stopped %s\n", name)))
 					}
 				}
-				if e := docker.DockerRemove(container.ID); e != nil {
+				if e := containers.Remove(container.ID); e != nil {
 					if !Service.HostConfig.AutoRemove {
 						color.Print(Green(fmt.Sprintf("Successfully removed %s\n", name)))
 					}
@@ -235,7 +236,7 @@ func (Service *Service) Stop() error {
 	}
 
 	for _, name := range container.Names {
-		if e := docker.DockerStop(container.ID); e == nil {
+		if e := containers.Stop(container.ID); e == nil {
 			if !discrete {
 				containerName := strings.Trim(name, "/")
 				color.Print(Green(fmt.Sprintf("Successfully stopped %v\n", containerName)))
@@ -264,8 +265,8 @@ func (Service *Service) StopAndRemove() error {
 	}
 
 	for _, name := range container.Names {
-		if e := docker.DockerStop(container.ID); e == nil {
-			if e := docker.DockerRemove(container.ID); e == nil {
+		if e := containers.Stop(container.ID); e == nil {
+			if e := containers.Remove(container.ID); e == nil {
 				if !discrete {
 					containerName := strings.Trim(name, "/")
 					fmt.Print(Green(fmt.Sprintf("Successfully removed %v\n", containerName)))
@@ -287,7 +288,7 @@ func (Service *Service) Remove() error {
 
 	for _, name := range container.Names {
 		containerName := strings.Trim(name, "/")
-		if e := docker.DockerRemove(container.ID); e == nil {
+		if e := containers.Remove(container.ID); e == nil {
 			if !discrete {
 				fmt.Print(Green(fmt.Sprintf("Successfully removed %s\n", containerName)))
 			}
@@ -312,7 +313,7 @@ func (Service *Service) DockerLogs() ([]byte, error) {
 	}
 
 	name, _ := Service.GetFieldString("name")
-	return docker.DockerContainerLogs(name)
+	return containers.Logs(name)
 }
 
 // DockerRun will start an existing container.
@@ -329,7 +330,7 @@ func (Service *Service) DockerRun() error {
 	if e != nil {
 		return fmt.Errorf("container config is missing label for name")
 	}
-	if err := docker.DockerContainerStart(name, container.StartOptions{}); err != nil {
+	if err := containers.Start(name, container.StartOptions{}); err != nil {
 		return err
 	}
 
@@ -352,7 +353,7 @@ func (Service *Service) DockerRunInteractive() error {
 		return fmt.Errorf("container config is missing label for name")
 	}
 
-	waiter, err := docker.DockerContainerAttach(name, container.AttachOptions{
+	waiter, err := containers.Attach(name, container.AttachOptions{
 		Stderr: true,
 		Stdout: true,
 		Stdin:  true,
@@ -381,7 +382,7 @@ func (Service *Service) DockerRunInteractive() error {
 		}
 	}()
 
-	if err := docker.DockerContainerStart(name, container.StartOptions{}); err != nil {
+	if err := containers.Start(name, container.StartOptions{}); err != nil {
 		return err
 	}
 
@@ -401,7 +402,7 @@ func (Service *Service) DockerRunInteractive() error {
 		}()
 	}
 
-	if err := docker.DockerContainerWait(name, container.WaitConditionNotRunning); err != nil {
+	if err := containers.Wait(name, container.WaitConditionNotRunning); err != nil {
 		return err
 	}
 
@@ -420,7 +421,7 @@ func (Service *Service) DockerCreate() error {
 	}
 
 	// Sanity check to ensure we don't get name conflicts.
-	c, _ := docker.DockerContainerList()
+	c, _ := containers.List()
 	for _, cn := range c {
 		if strings.HasSuffix(cn.Names[0], Service.Config.Labels["pygmy.name"]) {
 			return fmt.Errorf("container already created, or namespace is already taken")
@@ -432,7 +433,7 @@ func (Service *Service) DockerCreate() error {
 		return fmt.Errorf("container config is missing label for name")
 	}
 
-	_, err = docker.DockerContainerCreate(name, Service.Config, Service.HostConfig, Service.NetworkConfig)
+	_, err = containers.Create(name, Service.Config, Service.HostConfig, Service.NetworkConfig)
 	if err != nil {
 		return err
 	}
